+++ src/app/api/kamu.graphql.interface.ts
@@ -1984,6 +1984,39 @@ export type DatasetResumeFlowsMutation = {
     };
 };
 
+export type DatasetTriggerFlowMutationVariables = Exact<{
+    datasetId: Scalars["DatasetID"];
+    datasetFlowType: DatasetFlowType;
+}>;
+
+export type DatasetTriggerFlowMutation = {
+    __typename?: "Mutation";
+    datasets: {
+        __typename?: "DatasetsMut";
+        byId?: {
+            __typename?: "DatasetMut";
+            flows: {
+                __typename?: "DatasetFlowsMut";
+                runs: {
+                    __typename?: "DatasetFlowRunsMut";
+                    triggerFlow:
+                        | {
+                              __typename?: "FlowIncompatibleDatasetKind";
+                              expectedDatasetKind: DatasetKind;
+                              actualDatasetKind: DatasetKind;
+                              message: string;
+                          }
+                        | {
+                              __typename?: "TriggerFlowSuccess";
+                              message: string;
+                              flow: { __typename?: "Flow" } & FlowSummaryDataFragment;
+                          };
+                };
+            };
+        } | null;
+    };
+};
+
 export type FlowSummaryDataFragment = {
     __typename?: "Flow";
     flowId: string;
@@ -2026,7 +2059,7 @@ export type FlowSummaryDataFragment = {
               } | null;
           }
         | { __typename?: "FlowDescriptionSystemGC"; dummy: boolean };
-    initiator?: { __typename?: "Account"; accountName: string } | null;
+    initiator?: ({ __typename?: "Account" } & AccountFragment) | null;
     timing: {
         __typename?: "FlowTimingRecords";
         activateAt?: string | null;
@@ -2798,6 +2831,16 @@ export type SearchDatasetsOverviewQuery = {
     };
 };
 
+export const AccountFragmentDoc = gql`
+    fragment Account on Account {
+        id
+        accountName
+        displayName
+        accountType
+        avatarUrl
+        isAdmin
+    }
+`;
 export const FlowSummaryDataFragmentDoc = gql`
     fragment FlowSummaryData on Flow {
         description {
@@ -2836,7 +2879,7 @@ export const FlowSummaryDataFragmentDoc = gql`
         flowId
         status
         initiator {
-            accountName
+            ...Account
         }
         outcome
         timing {
@@ -2854,6 +2897,7 @@ export const FlowSummaryDataFragmentDoc = gql`
             }
         }
     }
+    ${AccountFragmentDoc}
 `;
 export const DatasetPageInfoFragmentDoc = gql`
     fragment DatasetPageInfo on PageBasedInfo {
@@ -2880,16 +2924,6 @@ export const FlowConnectionDataFragmentDoc = gql`
     }
     ${FlowSummaryDataFragmentDoc}
     ${DatasetPageInfoFragmentDoc}
-`;
-export const AccountFragmentDoc = gql`
-    fragment Account on Account {
-        id
-        accountName
-        displayName
-        accountType
-        avatarUrl
-        isAdmin
-    }
 `;
 export const DatasetDataSizeFragmentDoc = gql`
     fragment DatasetDataSize on DatasetData {
@@ -4257,6 +4291,46 @@ export class DatasetResumeFlowsGQL extends Apollo.Mutation<
     DatasetResumeFlowsMutationVariables
 > {
     document = DatasetResumeFlowsDocument;
+
+    constructor(apollo: Apollo.Apollo) {
+        super(apollo);
+    }
+}
+export const DatasetTriggerFlowDocument = gql`
+    mutation datasetTriggerFlow($datasetId: DatasetID!, $datasetFlowType: DatasetFlowType!) {
+        datasets {
+            byId(datasetId: $datasetId) {
+                flows {
+                    runs {
+                        triggerFlow(datasetFlowType: $datasetFlowType) {
+                            ... on TriggerFlowSuccess {
+                                flow {
+                                    ...FlowSummaryData
+                                }
+                                message
+                            }
+                            ... on FlowIncompatibleDatasetKind {
+                                expectedDatasetKind
+                                actualDatasetKind
+                                message
+                            }
+                        }
+                    }
+                }
+            }
+        }
+    }
+    ${FlowSummaryDataFragmentDoc}
+`;
+
+@Injectable({
+    providedIn: "root",
+})
+export class DatasetTriggerFlowGQL extends Apollo.Mutation<
+    DatasetTriggerFlowMutation,
+    DatasetTriggerFlowMutationVariables
+> {
+    document = DatasetTriggerFlowDocument;
 
     constructor(apollo: Apollo.Apollo) {
         super(apollo);